refactor

+++ core-common/src/main/java/eu/arrowhead/common/CommonConstants.java
@@ -106,7 +106,7 @@ public class CommonConstants {
 	public static final int SYSTEM_PORT_RANGE_MIN = 0;
 	public static final int SYSTEM_PORT_RANGE_MAX = 65535;
 	
-	public static final int CONVERSION_MILLISECOND_TO_MINUTES = 60000;
+	public static final long CONVERSION_MILLISECOND_TO_MINUTES = 60000;
 
 	private CommonConstants() {
 		throw new UnsupportedOperationException();
